requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
colorama==0.4.6
ecdsa==0.19.0
fastapi==0.110.1
fastapi-security==0.5.0
h11==0.14.0
idna==3.7
psycopg==3.1.18
psycopg-pool==3.2.1
pyasn1==0.6.0
pydantic==1.10.15
pydantic_core==2.18.1
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
tzdata==2024.1
uvicorn==0.29.0
